# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Buddy360 API",
    "description": "Backend API for the Buddy360 parenting-companion app. Provides authentication, child-profile management, LLM-powered guidance, and audio transcription.",
    "contact": {
      "name": "Buddy360 Engineering",
      "email": "prince.kumar@pearson.com"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current environment"
    }
  ],
  "paths": {
...
    "/api/v1/jobs": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Enqueue Job",
        "description": "Enqueue an LLM job. Returns a job_id immediately \u2014 the worker processes the job asynchronously. Poll GET /jobs/{job_id} for completion.",
        "operationId": "enqueue_job_api_v1_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnqueueJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnqueueJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job Status",
        "description": "Poll the status of a previously enqueued job. Re-fetch domain data when status == 'completed'.",
        "operationId": "get_job_status_api_v1_jobs__job_id__get",
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EnqueueJobPayload": {
        "properties": {
          "prompt": {
            "type": "string",
            "maxLength": 32000,
            "title": "Prompt"
          },
          "response_json_schema": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Json Schema"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          }
        },
        "type": "object",
        "required": [
          "prompt"
        ],
        "title": "EnqueueJobPayload"
      },
      "EnqueueJobRequest": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "generate_recommendations",
              "generate_activity",
              "generate_personality_analysis",
              "generate_life_pathway",
              "generate_growth_parent_questions",
              "generate_growth_child_rounds",
              "generate_observations"
            ],
            "title": "Type"
          },
          "child_id": {
            "type": "string",
            "maxLength": 100,
            "title": "Child Id"
          },
          "payload": {
            "$ref": "#/components/schemas/EnqueueJobPayload"
          },
          "write_back": {
            "$ref": "#/components/schemas/WriteBackConfig"
          }
        },
        "type": "object",
        "required": [
          "type",
          "child_id",
          "payload",
          "write_back"
        ],
        "title": "EnqueueJobRequest"
      },
      "EnqueueJobResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          }
        },
        "type": "object",
        "required": [
          "job_id"
        ],
        "title": "EnqueueJobResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobStatusResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status",
          "created_at"
        ],
        "title": "JobStatusResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WriteBackConfig": {
        "properties": {
          "collection": {
            "type": "string",
            "maxLength": 50,
            "title": "Collection"
          },
          "filter": {
            "additionalProperties": true,
            "type": "object",
            "title": "Filter"
          },
          "field": {
            "type": "string",
            "maxLength": 200,
            "title": "Field"
          }
        },
        "type": "object",
        "required": [
          "collection",
          "filter",
          "field"
        ],
        "title": "WriteBackConfig"
      }
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      },
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}